// web/lib/api/generated/model/checkoutInputRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Inventory Platform API
 * Multi-tenant B2B inventory & ordering platform. All money values are decimal strings.
 * OpenAPI spec version: 1.0.0
 */
import type { CheckoutPurposeEnum } from './checkoutPurposeEnum';

export interface CheckoutInputRequest {
  purpose: CheckoutPurposeEnum;
  /** @nullable */
  invoice_id?: string | null;
  /**
     * @nullable
     * @pattern ^-?\d{0,12}(?:\.\d{0,2})?$
     */
  amount?: string | null;
}