// web/lib/api/generated/model/checkoutOutcomeOutcomeEnum.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Inventory Platform API
 * Multi-tenant B2B inventory & ordering platform. All money values are decimal strings.
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `success` - success
 * * `failed` - failed
 * * `dismissed` - dismissed
 */
export type CheckoutOutcomeOutcomeEnum = typeof CheckoutOutcomeOutcomeEnum[keyof typeof CheckoutOutcomeOutcomeEnum];


export const CheckoutOutcomeOutcomeEnum = {
  success: 'success',
  failed: 'failed',
  dismissed: 'dismissed',
} as const;